Add support for streaming all users contained in a twitter public list
`user_lists` is a shortcut to follow all members of a public
twitter list identified by the user id and the list slug (last part of uri when open a list in your browser). The
configuration is the same as the twitter API (a single comma separated string value, or using json arrays).

```
PUT _river/my_twitter_river/_meta
{
    "type" : "twitter",
    "twitter" : {
        "filter" : {
            "tracks" : "test,something,please",
            "follow" : "111,222,333",
            "user_lists" : "ownerScreenName1/slug1,ownerScreenName2/slug2",
            "locations" : "-122.75,36.8,-121.75,37.8,-74,40,-73,41",
            "language" : "fr,en"
        }
    }
}
```

Closes #71.