Remove explicit op_type=create
When running multiple Twitter Rivers which fill the same ES index, there is a possibility that the rivers will fetch the same Twitter status with different tracking keywords. The Twitter River uses `op_type=create` which causes the bulk operation to report a warning: "There was failures while executing bulk".

Closes #70.