Put mapping only if river does not exist
By default, we always try to create a mapping even if it already exists.
We should be smarter and don't send to cluster a non needed update.

We also add more logs so we can have a better understanding when we have issues while starting river.

(cherry picked from commit 1d84190)