Elasticsearch 1.4.0 hangs on running twitter river
When running Twitter in elasticsearch 1.4, it could happen that some cluster API don't work anymore.
The river is not created properly so the cluster does not seem to be updated and is kept in a blocked state.

This patch makes the `start()` method returning as soon as possible and spawn a Thread which will actually do the initialization when the cluster is in yellow or green state.

Closes #83.

(cherry picked from commit fef2138)